Update dom updates file
